EXAM-29: Move common from modules
