+++ furniture-server/src/common/constants.ts
