+++ furniture-server/src/common/errorInterceptor.ts
