+++ furniture-server/src/modules/product/product.controller.ts
@@ -17,7 +17,7 @@ import { GetProductsQueryDto } from './dto/getProductsQuery.dto';
 import { Roles } from '../auth/roles-guard/roles.decorator';
 import { EUserRole } from '../../models/user/user.entity';
 import { RolesGuard } from '../auth/roles-guard/roles.guard';
-import { ErrorInterceptor } from '../common/errorInterceptor';
+import { ErrorInterceptor } from '../../common/errorInterceptor';
 
 
 @UseInterceptors(ErrorInterceptor)
